chore(directories): move stripe-properties to configuration directory
