+++ src/main/kotlin/techcourse/herobeans/HeroBeansApplication.kt
@@ -4,7 +4,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.context.properties.EnableConfigurationProperties
 import org.springframework.boot.runApplication
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing
-import techcourse.herobeans.config.StripeProperties
+import techcourse.herobeans.configuration.StripeProperties
 
 @EnableConfigurationProperties(StripeProperties::class)
 @SpringBootApplication
